perf(tests): cut the CI test critical path
Split the 5-test wall-time zone sweep into one file per test so vitest can
spread it across workers, drop the Map allocation from instantToWallTime,
make the polling sleeps honour faked timers, shrink the live-driver lifetime
waits, and shard the CI test job three ways.
