+++ .github/workflows/checks.yml
@@ -30,6 +30,16 @@ jobs:
 
   test:
     runs-on: ubuntu-latest
+    strategy:
+      fail-fast: false
+      matrix:
+        include:
+          - shard: calendar
+            filter: --filter @keeper.sh/calendar
+          - shard: services
+            filter: --filter @keeper.sh/api --filter @keeper.sh/cron
+          - shard: rest
+            filter: --filter '!@keeper.sh/calendar' --filter '!@keeper.sh/api' --filter '!@keeper.sh/cron'
     services:
       postgres:
         image: postgres@sha256:ef257d85f76e48da1c64832459b59fcaba1a4dac97bf5d7450c77753542eee94 # 17.6-alpine
@@ -46,7 +56,8 @@ jobs:
       - uses: actions/checkout@3d3c42e5aac5ba805825da76410c181273ba90b1 # v7.0.1
       - uses: oven-sh/setup-bun@0c5077e51419868618aeaa5fe8019c62421857d6 # v2
       - run: bun install --frozen-lockfile
-      - run: bun run test
+      - run: bun -e 'console.log(require("node:os").availableParallelism(), require("node:os").cpus().length)'
+      - run: bunx turbo run test ${{ matrix.filter }}
         env:
           KEEPER_TEST_DATABASE_URL: postgres://postgres:postgres@localhost:5432/postgres
