+++ services/api/vitest.config.ts
@@ -13,8 +13,8 @@ export default defineConfig({
     /*
      * Suites here boot an in-process PGlite and apply their schema in a setup
      * hook. That costs well under a second on a developer machine, but CI runs
-     * every workspace's vitest concurrently on a two-core runner alongside a
-     * Postgres service container, and the default ten seconds is not enough
+     * several workspaces' vitest concurrently on a small shared runner alongside
+     * a Postgres service container, and the default ten seconds is not enough
      * headroom for it there.
      */
     hookTimeout: 60_000,
